Multiple small changes as identified by Andras Mohari:
(c) should be \copyright in the copyright message.

5.3.1 Setting Your Location: \" should be " in the config file
excerpt.

5.3.2 Setting the Display Resolution: LyX-code should be used
(instead of Quote) for the config file excerpt.

5.6 Other .fab files: The * in "Note that * in..." should be
typeset using typewriter font so that it matches the *.fab text in
the same paragraph.

A. Configuration File: All occurences of colour should be
replaced with color in section names and IDs.  (That's the spelling
used in the config file.)

2.4 Running Stellarium:
Windows: The Stellarium installer creates an item in the Start Menu under in Programs section. Select this to run Stellarium.

5.3 The Main Configuration File:
Windows: The configuration file, config.ini, is in the config sub-directory of the main Stellarium directory. If you installed to the default location, this will be C:\Program Files\Stellarium\config.

* A. Configuration File: init_view_pos. should read:

initial viewing direction. This is a vector with x,y,z-coordinates. x
being N-S (S +ve), y being E-W (E +ve), z being up-down (up +ve).
Thus to look South at the horizon use 0,1,0. To look Northwest and up
at 45 degrees, use -1,-1,1 and so on.

A. Configuration File: flag_milky_intensity (float): should be
"milky_way_intensity".